+++ tests/test_client_common.py
@@ -5,6 +5,7 @@
 
 from frequenz.client.common.microgrid.components import (
     ComponentCategory,
+    ComponentErrorCode,
     ComponentStateCode,
 )
 
@@ -19,3 +20,9 @@ def test_component_state_code() -> None:
     """Test the component state code."""
     for state_code in ComponentStateCode:
         assert ComponentStateCode.from_proto(state_code.to_proto()) == state_code
+
+
+def test_component_error_code() -> None:
+    """Test the component error code."""
+    for error_code in ComponentErrorCode:
+        assert ComponentErrorCode.from_proto(error_code.to_proto()) == error_code